Support ALL AccuDraw shortcuts when modifying edit handles.
Currently because both EditManipulator.HandleTool and interactive
AccuDraw shortcuts are InputCollector sub-classes you are unable to
use Rotate Axes/Element when dragging a handle.

Now when an InputCollector is active, these shortcuts install as
a ViewTool, which suspends the InputCollector.

Moved some classes into their own files to reduce introducing circular
dependencies.
