+++ core/frontend/src/core-frontend.ts
@@ -169,6 +169,7 @@ export * from "./tools/IdleTool";
 export * from "./tools/MeasureTool";
 export * from "./tools/PrimitiveTool";
 export * from "./tools/SelectTool";
+export * from "./tools/SetupCameraTools";
 export * from "./tools/Tool";
 export * from "./tools/ToolSettings";
 export * from "./tools/ToolAdmin";
